Ch7 Step Defs from inside Separating support code
pg136
Created a support class java file in src/test/support called
KnowsTheDomain and added the class.  Import statements needed adding in
KnowsTheDomain class, and an import status was added to the steps file
(used the eclipse help for this)
